Treat long type as float in Tableau to adapt to an index type error

+++ connector/elasticsearch-connector.js
@@ -5,7 +5,7 @@ var elasticsearchConnector = (function () {
         keyword:  tableau.dataTypeEnum.string,
         text:  tableau.dataTypeEnum.string,
         float:  tableau.dataTypeEnum.float,
-        long:  tableau.dataTypeEnum.int,
+        long:  tableau.dataTypeEnum.float,
         integer:  tableau.dataTypeEnum.int,
         double:  tableau.dataTypeEnum.float,
         date:  tableau.dataTypeEnum.datetime,
